remove deprecated `header` params from HttpClient
